[form-builder] Wrap asset source plugin components in withDocument instead of the image input
Previously the ImageInput was wrapped by the `withDocument` HOC that provided it with the current document as a prop. This was then passed to the asset source plugin for convenience. This meant that whenever the document changed the image input would then receive a new `document` prop. Since it only forwarded the document to the asset source plugin, this patch instead wraps the asset source plugin in `withDocument`.

This should improve the render performance of the ImageInput since it no longer receives the enclosing `document` as a prop, which changes on every keystroke.